change abstracts and implementation
