+++ packages/dnd/src/Draggable/DraggableAxes.ts
@@ -1,5 +1,5 @@
 import { AbstractDraggable } from "@dflex/draggable";
-import type { DraggedStyle, Coordinates } from "@dflex/draggable";
+import type { Coordinates } from "@dflex/draggable";
 
 import { AxesCoordinates, AxesCoordinatesBool } from "@dflex/utils";
 import type {
@@ -11,35 +11,31 @@ import type { CoreInstanceInterface } from "@dflex/core-instance";
 
 import store from "../DnDStore";
 
-import type { DraggableDnDInterface, Restrictions } from "./types";
-
 import type {
-  ScrollOptWithThreshold,
-  FinalDndOpts,
-  RestrictionsStatus,
-} from "../types";
+  DraggableAxesInterface,
+  SiblingsThreshold,
+  Restrictions,
+} from "./types";
 
-import Threshold, { ThresholdCoordinate } from "../Plugins/Threshold";
+import type { FinalDndOpts, RestrictionsStatus } from "../types";
 
-class Draggable
+import Threshold from "../Plugins/Threshold";
+import type {
+  ThresholdInterface,
+  ThresholdCoordinate,
+} from "../Plugins/Threshold";
+
+class DraggableAxes
   extends AbstractDraggable<CoreInstanceInterface>
-  implements DraggableDnDInterface
+  implements DraggableAxesInterface
 {
   private isLayoutStateUpdated: boolean;
 
   indexPlaceholder: number;
 
-  operationID: string;
+  threshold: ThresholdInterface;
 
-  siblingsContainer: CoreInstanceInterface | null;
-
-  setOfTransformedIds?: Set<string>;
-
-  threshold: DraggableDnDInterface["threshold"];
-
-  layoutThresholds!: DraggableDnDInterface["layoutThresholds"];
-
-  scroll: ScrollOptWithThreshold;
+  layoutThresholds: SiblingsThreshold;
 
   isViewportRestricted: boolean;
 
@@ -47,15 +43,9 @@ class Draggable
 
   offsetPlaceholder: AxesCoordinatesInterface;
 
-  occupiedOffset: AxesCoordinatesInterface;
-
-  occupiedTranslate: AxesCoordinatesInterface;
-
   mousePoints: AxesCoordinatesInterface;
 
   isMovingAwayFrom: AxesCoordinatesBoolInterface;
-
-  numberOfElementsTransformed: number;
 
   isDraggedOutPosition: AxesCoordinatesBoolInterface;
 
@@ -73,12 +63,8 @@ class Draggable
 
   private initX: number;
 
-  isDraggedPositionFixed: boolean;
-
-  private changeToFixedStyleProps: DraggedStyle;
-
   constructor(id: string, initCoordinates: Coordinates, opts: FinalDndOpts) {
-    const { element, parent } = store.getElmTreeById(id);
+    const { element } = store.getElmTreeById(id);
 
     super(element, initCoordinates);
 
@@ -92,62 +78,17 @@ class Draggable
      */
     this.indexPlaceholder = order.self;
 
-    // This tiny bug caused an override  options despite it's actually freezed!
-    this.scroll = { ...opts.scroll };
-
     const { SK } = store.registry[id].keys;
-
-    const { hasOverflowX, hasOverflowY } = store.siblingsScrollElement[SK];
 
     const siblings = store.getElmSiblingsListById(this.draggedElm.id);
 
     this.isViewportRestricted = true;
-    this.isDraggedPositionFixed = false;
-
-    this.changeToFixedStyleProps = [
-      {
-        prop: "top",
-        dragValue: `${this.draggedElm.currentPosition.y}px`,
-        afterDragValue: "",
-      },
-      {
-        prop: "left",
-        dragValue: `${this.draggedElm.currentPosition.x}px`,
-        afterDragValue: "",
-      },
-      {
-        prop: "position",
-        dragValue: "fixed",
-        afterDragValue: "",
-      },
-    ];
-
-    if (siblings === null || (!hasOverflowY && !hasOverflowX)) {
-      // Override the default options. (FYI, this is the only privilege I have.)
-      this.scroll.enable = false;
-    }
-
-    if (this.scroll.enable) {
-      this.isViewportRestricted = false;
-
-      store.siblingsScrollElement[SK].setThresholdMatrix(this.scroll.threshold);
-
-      if (!store.siblingsScrollElement[SK].hasDocumentAsContainer) {
-        /**
-         * When the scroll is the document it's good. The restriction is to the
-         * document which guarantees the free movement. Otherwise, let's do it.
-         * Change the position and transform siblings.
-         */
-        // this.isDraggedPositionFixed = true;
-      }
-    }
 
     const siblingsBoundaries = store.siblingsBoundaries[SK];
 
     const {
       offset: { width, height },
       currentPosition,
-      translate,
     } = this.draggedElm;
 
     this.threshold = new Threshold(
@@ -177,24 +118,13 @@ class Draggable
           true
         ),
       };
+    } else {
+      this.layoutThresholds = {};
     }
-
-    this.siblingsContainer = null;
 
     this.isDraggedOutPosition = new AxesCoordinatesBool(false, false);
     this.isDraggedOutContainer = new AxesCoordinatesBool(false, false);
     this.isMovingAwayFrom = new AxesCoordinatesBool(false, false);
-
-    if (parent) {
-      /**
-       * Indicator to parents that have changed. This facilitates looping in
-       * affected parents only.
-       */
-      this.setOfTransformedIds = new Set([]);
-      this.assignActiveParent(parent);
-    }
-
-    this.operationID = store.tracker.newTravel();
 
     const { x, y } = initCoordinates;
 
@@ -218,20 +148,7 @@ class Draggable
       currentPosition.y
     );
 
-    this.occupiedOffset = new AxesCoordinates(
-      currentPosition.x,
-      currentPosition.y
-    );
-
-    this.occupiedTranslate = new AxesCoordinates(translate.x, translate.y);
-
     this.mousePoints = new AxesCoordinates(x, y);
-
-    /**
-     * It counts number of element that dragged has passed. This counter is
-     * crucial to calculate drag's translate and index
-     */
-    this.numberOfElementsTransformed = 0;
 
     this.restrictions = opts.restrictions;
 
@@ -241,25 +158,6 @@ class Draggable
       siblings !== null &&
       (opts.restrictionsStatus.isContainerRestricted ||
         opts.restrictionsStatus.isSelfRestricted);
-  }
-
-  /**
-   * Assigns new container parent to the dragged.
-   *
-   * @param element -
-   */
-  private assignActiveParent(element: CoreInstanceInterface) {
-    /**
-     * Assign a new instance which represents droppable. Then
-     * assign owner parent so we have from/to.
-     */
-    this.siblingsContainer = element;
-
-    /**
-     * Add flag for undo method so we can check which  parent is being
-     * transformed and which is not.
-     */
-    this.isDraggedOutContainer.setFalsy();
   }
 
   private axesYFilter(
@@ -317,9 +215,13 @@ class Draggable
     return x;
   }
 
-  setDraggedTempIndex(i: number) {
-    this.indexPlaceholder = i;
-    this.draggedElm.setDataset("index", i);
+  private setDraggedMouseMovement(x: number, y: number) {
+    this.isMovingAwayFrom.setAxes(
+      x > this.mousePoints.x,
+      y > this.mousePoints.y
+    );
+
+    this.mousePoints.setAxes(x, y);
   }
 
   /**
@@ -340,6 +242,8 @@ class Draggable
       this.isLayoutStateUpdated = true;
       store.onStateChange("dragging");
     }
+
+    this.setDraggedMouseMovement(x, y);
 
     let filteredY = y;
     let filteredX = x;
@@ -469,7 +373,7 @@ class Draggable
 
   isLeavingFromHead() {
     return (
-      this.isMovingAwayFrom.x && this.indexPlaceholder <= 0 // first our outside.
+      this.isMovingAwayFrom.y && this.indexPlaceholder <= 0 // first our outside.
     );
   }
 
@@ -477,7 +381,7 @@ class Draggable
     const lastElm =
       (store.getElmSiblingsListById(this.draggedElm.id) as string[]).length - 1;
 
-    return this.isMovingAwayFrom.y && this.indexPlaceholder === lastElm;
+    return !this.isMovingAwayFrom.y && this.indexPlaceholder === lastElm;
   }
 
   isNotSettled() {
@@ -488,85 +392,6 @@ class Draggable
       (this.isOutThreshold() || this.isOutThreshold(SK))
     );
   }
-
-  setDraggedMouseMovement(x: number, y: number) {
-    this.isMovingAwayFrom.setAxes(
-      x > this.mousePoints.x,
-      y > this.mousePoints.y
-    );
-
-    this.mousePoints.setAxes(x, y);
-  }
-
-  updateNumOfElementsTransformed(effectedElemDirection: number) {
-    this.numberOfElementsTransformed += -1 * effectedElemDirection;
-  }
-
-  setDraggedTransformPosition(isFallback: boolean) {
-    const siblings = store.getElmSiblingsListById(this.draggedElm.id);
-
-    /**
-     * In this case, the use clicked without making any move.
-     */
-    if (
-      isFallback ||
-      siblings === null ||
-      this.numberOfElementsTransformed === 0
-    ) {
-      /**
-       * If not isDraggedOutPosition, it means dragged is out its position, inside
-       * list but didn't reach another element to replace.
-       *
-       * List's elements is in their position, just undo dragged.
-       *
-       * Restore dragged position (translateX, translateY) directly. Why? Because,
-       * dragged depends on extra instance to float in layout that is not related to element
-       * instance.
-       */
-      if (!this.draggedElm.translate.isEqual(this.translatePlaceholder)) {
-        this.draggedElm.transformElm();
-        this.draggedElm.setDataset("index", this.draggedElm.order.self);
-
-        /**
-         * There's a rare case where dragged leaves and returns to the same
-         * position. In this case, undo won't be triggered so that we have to do
-         * it manually here. Otherwise, undoing will handle repositioning. I
-         * don't like it but it is what it is.
-         */
-        if (
-          siblings &&
-          siblings[this.draggedElm.order.self] !== this.draggedElm.id
-        ) {
-          this.draggedElm.assignNewPosition(
-            siblings,
-            this.draggedElm.order.self
-          );
-        }
-      }
-
-      return;
-    }
-
-    this.draggedElm.currentPosition.clone(this.occupiedOffset);
-    this.draggedElm.translate.clone(this.occupiedTranslate);
-
-    this.draggedElm.transformElm();
-
-    if (siblings) {
-      this.draggedElm.assignNewPosition(siblings, this.indexPlaceholder);
-    }
-
-    this.draggedElm.order.self = this.indexPlaceholder;
-  }
-
-  endDragging(isFallback: boolean) {
-    this.setDragged(false);
-    this.setDraggedTransformPosition(isFallback);
-
-    if (this.isDraggedPositionFixed) {
-      this.changeStyle(this.changeToFixedStyleProps, false);
-    }
-  }
 }
 
-export default Draggable;
+export default DraggableAxes;
